recent-servers: Expose functions for UI
For the purpose of these functions, there is no need to set up a file
monitor, nor for the NautilusRecentServers object at all. But they are
provided here to keep context and reuse server_list_load().

The network address bar is going to be adding servers to the list,
while the upcoming network view is going to allow to remove them.