directory-async: Split out nautilus-request
Requests don't necessarily require a directory, so there is no reason
for them to live inside nautilus-directory. Also, moving them to a
separate file allows using them outside of nautilus-directory-async.
This is is done for vfs-file's ready check, which now does not have to
go through a redundant directory method anymore.

This is in preparation of separating out the callback code from
directory-async.